src/main/java/org/czareg/codewars/ones/and/zeros/BinaryArrayToNumber.java
package org.czareg.codewars.ones.and.zeros;

import lombok.experimental.UtilityClass;

import java.util.List;

/*
Given an array of ones and zeroes, convert the equivalent binary value to an integer.

Eg: [0, 0, 0, 1] is treated as 0001 which is the binary representation of 1.

Examples:

Testing: [0, 0, 0, 1] ==> 1
Testing: [0, 0, 1, 0] ==> 2
Testing: [0, 1, 0, 1] ==> 5
Testing: [1, 0, 0, 1] ==> 9
Testing: [0, 0, 1, 0] ==> 2
Testing: [0, 1, 1, 0] ==> 6
Testing: [1, 1, 1, 1] ==> 15
Testing: [1, 0, 1, 1] ==> 11
However, the arrays can have varying lengths, not just limited to 4.
 */
@UtilityClass
public class BinaryArrayToNumber {

    public static int convertBinaryArrayToInt(List<Integer> binary) {
        String binaryAsString = binary.stream()
                .collect(StringBuilder::new, StringBuilder::append, StringBuilder::append)
                .toString();
        return Integer.parseInt(binaryAsString, 2);
    }
}
